# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Manajemen Karyawan API",
        "description": "Dokumentasi API untuk sistem manajemen karyawan, absensi, dan penggajian",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/gaji-bulanan": {
            "get": {
                "tags": [
                    "Gaji Bulanan"
                ],
                "summary": "Mendapatkan semua data gaji bulanan",
                "operationId": "4ea27729d12665472d1dd8ea4426c0cc",
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GajiBulanan"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Gaji Bulanan"
                ],
                "summary": "Membuat data gaji bulanan baru",
                "operationId": "e7b777405b25c97972c3296f825aba5c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_absensi": {
                                        "type": "string",
                                        "example": "ABS-001"
                                    },
                                    "id_pembayaran": {
                                        "type": "string",
                                        "example": "PMB-001"
                                    },
                                    "nominal": {
                                        "type": "integer",
                                        "example": 5000000
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-23"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GajiBulanan"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/gaji-bulanan/{id}": {
            "get": {
                "tags": [
                    "Gaji Bulanan"
                ],
                "summary": "Mendapatkan data gaji bulanan berdasarkan ID",
                "operationId": "6780d3c81467113fdd50002280a9ccc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID gaji bulanan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GajiBulanan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Gaji Bulanan"
                ],
                "summary": "Memperbarui data gaji bulanan",
                "operationId": "0ac9ccf32804e5ed8870961202ece0f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID gaji bulanan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_absensi": {
                                        "type": "string",
                                        "example": "ABS-001"
                                    },
                                    "id_pembayaran": {
                                        "type": "string",
                                        "example": "PMB-001"
                                    },
                                    "nominal": {
                                        "type": "integer",
                                        "example": 5000000
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-23"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GajiBulanan"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Gaji Bulanan"
                ],
                "summary": "Menghapus data gaji bulanan",
                "operationId": "fb2490dc4927819eda340ca40b1f7a68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID gaji bulanan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/gaji-bulanan/karyawan/{id_karyawan}": {
            "get": {
                "tags": [
                    "Gaji Bulanan"
                ],
                "summary": "Mendapatkan gaji bulanan berdasarkan ID karyawan",
                "operationId": "18bd2b0b39abdc55c619297703d02a37",
                "parameters": [
                    {
                        "name": "id_karyawan",
                        "in": "path",
                        "description": "ID karyawan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GajiBulanan"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/gaji-bulanan/periode/{tahun}/{bulan}": {
            "get": {
                "tags": [
                    "Gaji Bulanan"
                ],
                "summary": "Mendapatkan gaji bulanan berdasarkan periode",
                "operationId": "f9d4c04c560cb0ed49a4f3b3ce5dfad5",
                "parameters": [
                    {
                        "name": "tahun",
                        "in": "path",
                        "description": "Tahun",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    },
                    {
                        "name": "bulan",
                        "in": "path",
                        "description": "Bulan (1-12)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GajiBulanan"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/gaji-bulanan/absensi/{id_absensi}": {
            "get": {
                "tags": [
                    "Gaji Bulanan"
                ],
                "summary": "Mendapatkan gaji bulanan berdasarkan ID absensi",
                "operationId": "7406151f0d89eb774e6fcf05439fbac5",
                "parameters": [
                    {
                        "name": "id_absensi",
                        "in": "path",
                        "description": "ID absensi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GajiBulanan"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/gaji-bulanan/pembayaran/{id_pembayaran}": {
            "get": {
                "tags": [
                    "Gaji Bulanan"
                ],
                "summary": "Mendapatkan gaji bulanan berdasarkan ID pembayaran",
                "operationId": "183c022e805aa31010916ecee04e764e",
                "parameters": [
                    {
                        "name": "id_pembayaran",
                        "in": "path",
                        "description": "ID pembayaran",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GajiBulanan"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GajiBulanan": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "id_absensi": {
                        "type": "string",
                        "example": "ABS-001"
                    },
                    "id_pembayaran": {
                        "type": "string",
                        "example": "PMB-001"
                    },
                    "nominal": {
                        "type": "integer",
                        "example": 5000000
                    },
                    "tanggal": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-04-23"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan API key dengan format Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}